LPD-54772 Move global types up
